Fix for PID file creation issue with Mosquitto 2.x.
Mosquitto 2.x defaults to writing a pid file in (/var)/run/mosquitto,
but that folder doesn't exist after boot and Mosquitto can't create
the folder due to permissions of the parent folder.

This change just creates (as root) the necessary folders.

I've also added a few variables for the systemd unit file due to
reasons.